+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### [Unreleased][unreleased]
 
+#### Changed
+- add support for different operators in `$conditions` array
+
 ### [1.8.0] - 2021-06-01
 #### Added
 - year column type (year for mysql, numeric(4) for pgsql)